Organize components into presentational and container folders
